clinic_ai_backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
pytest
ruff
python-multipart==0.0.9
openai-whisper
azure-storage-queue
